Windows: Changed commit hook linting to ignore empty diffs.
On Windows, if you attempt to commit an empty changeset, the commit hooks will
attempt to lint the entire repository. This is because we pass blank arguments
to the call to `xargs` that kicks off the C++ linter (e.g., in
`support/pre-commit`). In the git bash, the default behavior of blank arguments
is *not* to ignore them, as it is on certain other platforms.

Thus, we pass the `-r` flag in explicitly to `xargs`, which causes us to ignore
whitespace-only calls to `xargs`. This will remove the whole-repository linting
problem altogether.